Add short photo share URLs
