+++ src/app/p/[photoId]/not-found.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/app/share/photo/[photoId]/not-found';
