+++ src/app/p/[photoId]/opengraph-image/route.ts
@@ -1,0 +1,3 @@
+export { GET } from '@/app/share/photo/[photoId]/opengraph-image/route';
+
+export const runtime = 'nodejs';
